+++ ruoyi-admin/src/main/java/com/ruoyi/device/mapper/DeviceInfoMapper.java
@@ -5,15 +5,15 @@ import com.ruoyi.device.domain.DeviceInfo;
 
 /**
  * deviceInfoMapper接口
- * 
+ *
  * @author 王涛
  * @date 2021-05-07
  */
-public interface DeviceInfoMapper 
+public interface DeviceInfoMapper
 {
     /**
      * 查询deviceInfo
-     * 
+     *
      * @param id deviceInfoID
      * @return deviceInfo
      */
@@ -21,7 +21,7 @@ public interface DeviceInfoMapper
 
     /**
      * 查询deviceInfo列表
-     * 
+     *
      * @param deviceInfo deviceInfo
      * @return deviceInfo集合
      */
@@ -29,7 +29,7 @@ public interface DeviceInfoMapper
 
     /**
      * 新增deviceInfo
-     * 
+     *
      * @param deviceInfo deviceInfo
      * @return 结果
      */
@@ -37,7 +37,7 @@ public interface DeviceInfoMapper
 
     /**
      * 修改deviceInfo
-     * 
+     *
      * @param deviceInfo deviceInfo
      * @return 结果
      */
@@ -45,15 +45,15 @@ public interface DeviceInfoMapper
 
     /**
      * 删除deviceInfo
-     * 
+     *
      * @param id deviceInfoID
      * @return 结果
      */
-    public int deleteDeviceInfoById(Integer id);
+    public int deleteDeviceInfoById(Long id);
 
     /**
      * 批量删除deviceInfo
-     * 
+     *
      * @param ids 需要删除的数据ID
      * @return 结果
      */
